155 change npint to int

+++ _archive/blech_process_func.py
@@ -140,7 +140,7 @@ def blech_process(electrode_num):
                                 windows_in_data*(window_len * sampling_rate),
                                 windows_in_data)
     window_markers = np.array([int(x) for x in window_markers])
-    chosen_window_inds = np.vectorize(np.int)(np.sort(np.random.choice(
+    chosen_window_inds = np.vectorize(int)(np.sort(np.random.choice(
                                 np.arange(windows_in_data), window_count)))
     chosen_window_markers = [(window_markers[x-1],window_markers[x]) \
                                 for x in chosen_window_inds]
